Reorganize network context files
