--- src/context/actions.js
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) [2020] SUSE LLC
- *
- * All Rights Reserved.
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of version 2 of the GNU General Public License as published
- * by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, contact SUSE LLC.
- *
- * To contact SUSE LLC about this file by physical or electronic mail, you may
- * find current contact information at www.suse.com.
- */
-
-const SET_INTERFACES = 'set_interfaces';
-const SET_CONNECTIONS = 'set_connections';
-const SET_ROUTES = 'set_routes';
-const SET_DNS = 'set_dns';
-const UPDATE_ROUTES = 'update_routes';
-const ADD_CONNECTION = 'add_connection';
-const DELETE_CONNECTION = 'delete_connection';
-const UPDATE_CONNECTION = 'update_connection';
-const UPDATE_INTERFACE = 'update_interface';
-const CONNECTION_ERROR = 'connection_error';
-
-export {
-    SET_INTERFACES,
-    SET_CONNECTIONS,
-    SET_DNS,
-    SET_ROUTES,
-    UPDATE_ROUTES,
-    ADD_CONNECTION,
-    DELETE_CONNECTION,
-    UPDATE_CONNECTION,
-    UPDATE_INTERFACE,
-    CONNECTION_ERROR
-};
